+++ edge-util/src/Microsoft.Azure.Devices.Edge.Util/metrics/nullMetrics/NullMetricsGauge.cs
@@ -3,7 +3,7 @@ namespace Microsoft.Azure.Devices.Edge.Util.Metrics.NullMetrics
 {
     public class NullMetricsGauge : IMetricsGauge
     {
-        public void Set(long value, string[] labelValues)
+        public void Set(double value, string[] labelValues)
         {
         }
     }
