+++ edge-util/src/Microsoft.Azure.Devices.Edge.Util/metrics/prometheus.net/MetricsHistogram.cs
@@ -29,7 +29,7 @@ namespace Microsoft.Azure.Devices.Edge.Util.Metrics.Prometheus.Net
                 });
         }
 
-        public void Update(long value, string[] labelValues) =>
+        public void Update(double value, string[] labelValues) =>
             this.summary
                 .WithLabels(this.GetLabelValues(labelValues))
                 .Observe(value);
